Refactor store setup and extract router
